cfa-prep-tool/backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
pydantic==2.5.0
python-multipart==0.0.6
PyMuPDF==1.23.8
anthropic==0.7.8
python-dotenv==1.0.0
aiosqlite==0.19.0
